app/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
asyncpg>=0.28.0
google-genai>=0.2.0
python-dotenv>=1.0.0
